fix: add automatic versioning and release builds to ci
This modifies the release step so that we get automatic
version bumps and releases when merging to `main`. Releases
will still be created as a "draft" in GitHub, so we can control
when they get pulled into the registry.

This does not include any provisions for handling pre-releases.
Instead, that can be accomplished by manually downloading the
appropriate binary from the draft release.