fix(stream_exclusion_rule): set `title` property to optional
The exclusion rules created for Event Streaming configurations al-
ready have IDs which can be used as reference properties. To match
the behavor in our webapp, we opted to update the schemas on API +
TF to set the `title` property to optional. In the UI, the title
can be omitted which will still result in the creation of a valid
exclusion rule. This commit aligns the TF behavior with web along
with some minor housecleaning changes to ensure the coverage dir
is created before running tests + renames categories in test views

Semver: patch
Ref: LOG-10750