feat(alert&view): add support for integration via Slack
Defines the `slack_channel` argument in the schemas for resources
`logdna_alert` + `logdna_view` and the data source `logdna_alert`.
This facilitates the ability to integrate preset alerts and view-
specific alerts with Slack. Updated the tests accordingly and did
quite a bit of housecleaning to modularize & dedupe some code.

Ref: LOG-10589